Refactor code formatting in utils folder
The changes in the submitted code reflect alterations in formatting to improve readability. The major changes include adding a space between type and variable in function parameters according to the Google C++ style guide, altering the placement of ampersands for readability and consistency, and aligning indentation for better code clarity. These changes have not altered the underlying logic of the code.